Backend terminado

+++ database/migrations/2026_03_04_143056_create_objetos_table.php
@@ -9,17 +9,17 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-public function up(): void
-{
-    Schema::create('objetos', function (Blueprint $table) {
-        $table->id();
-        $table->string('nombre');
-        $table->text('descripcion')->nullable();
-        $table->decimal('precio', 8, 2);
-        $table->string('imagen')->nullable();
-        $table->timestamps();
-    });
-}
+    public function up(): void
+    {
+        Schema::create('objetos', function (Blueprint $table) {
+            $table->id();
+            $table->string('nombre');
+            $table->text('descripcion')->nullable();
+            $table->integer('precio');
+            $table->string('imagen')->nullable();
+            $table->timestamps();
+        });
+    }
 
     /**
      * Reverse the migrations.
